renaming unit-test directories to be discoverable
