--- Logging/Unit-Tests/test_Logger.log
@@ -1,3 +1,0 @@
-2023-09-20 10:15:15,072 | [INFO] 1stTest
-2023-09-20 10:15:15,072 | [INFO] 2ndTest
-2023-09-20 10:15:15,072 | [WARNING] 3rdTest
